feat: divide common contentView into two classes
- ScrollableContentViewWithHead
- ContentViewWithHead

#14, #15
